pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smol"
version = "0.1.0"
description = "A database of small connected graphs with their spectral properties"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Leo Torres", email = "leo@leotrs.com"}
]
keywords = ["graphs", "spectral", "eigenvalues", "cospectral", "network science"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "networkx>=3.0",
    "psycopg2-binary>=2.9",
    "fastapi>=0.124.4",
    "uvicorn>=0.38.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "matplotlib>=3.10.8",
    "aiosqlite>=0.22.0",
    "sympy>=1.14.0",
    "pot>=0.9.6.post1",
]

[project.urls]
Homepage = "https://github.com/leotrs/smol"
Repository = "https://github.com/leotrs/smol"
Issues = "https://github.com/leotrs/smol/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
include = ["db*"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "nbconvert>=7.16.6",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]

[tool.pytest.ini_options]
markers = [
    "needs_db: marks tests that require a populated database (skip in CI)",
]

[tool.ruff]
exclude = [
    "scripts/algebraic_approach.py",
    "scripts/analyze_*.py",
    "scripts/block_structure_analysis.py",
    "scripts/boundary_analysis.py",
    "scripts/check_*.py",
    "scripts/classify_switches.py",
    "scripts/combined_transfer_analysis.py",
    "scripts/compute_all_resolvents.py",
    "scripts/deep_*.py",
    "scripts/eigenvalue_analysis.py",
    "scripts/final_verification.py",
    "scripts/find_*.py",
    "scripts/fix_eigenvalue_sort.py",
    "scripts/investigate_*.py",
    "scripts/lumped_trace_proof.py",
    "scripts/matrix_difference_analysis.py",
    "scripts/mechanism_*.py",
    "scripts/new_nbl_analysis.py",
    "scripts/prove_*.py",
    "scripts/research_nbl.py",
    "scripts/rigorous_*.py",
    "scripts/structural_analysis.py",
    "scripts/switch_*.py",
    "scripts/test_*.py",
    "scripts/trace_decomposition.py",
    "scripts/verify_*.py",
    # Root-level research scripts (temporary, not part of production codebase)
    "analyze_*.py",
    "check_*.py",
    "compare_*.py",
    "deep_*.py",
    "examine_*.py",
    "find_*.py",
    "full_*.py",
    "pattern_*.py",
    "search_*.py",
    "test_*.py",
    "verify_*.py",
]
